+++ src/main/java/thelm/packagedauto/integration/emi/EncoderDragDropHandler.java
@@ -84,7 +84,7 @@ public class EncoderDragDropHandler implements EmiDragDropHandler<EncoderScreen>
 		public void accept(EmiIngredient emiIngredient) {
 			ItemStack stack = wrapStack(emiIngredient);
 			if(!stack.isEmpty()) {
-				PacketDistributor.sendToServer(new SetItemStackPacket((short)slot.index, stack));
+				PacketDistributor.sendToServer(new SetItemStackPacket(slot.index, stack));
 			}
 		}
 	}
